chore: Refactor Plot attribute assignment.
